// src/@types/UpdateStatus.d.ts
export type updatestatusfieldsProps={
    leadStatus:number,
    competitor:string,
    competitor_id:string,
    comment:string,
    competitorName:string,
    enquiry_type:string,
    demo_date:string,
    poc_date:string,
    follow_up_date:string,
}

export type leadStatusdropdownProps={
    leadStatusId:string,
    leadStatusName:string
}
export type competitordropdownProps={
    competitorId:string,
    competitorName:string
}
export type enquirydropdownProps={
    enquiryId:string,
    enquiryName:string
}